LayoutTests/media/video-controls-start-selection.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<p>Start selection on video with controls</p>
<video controls></video>
<span>Some text here</span>
<script>
    var video = document.querySelector('video');
    var span = document.querySelector('span');
    if (window.eventSender) {
        eventSender.mouseMoveTo(video.offsetLeft + video.offsetWidth / 2, video.offsetTop + video.offsetHeight / 2);
        eventSender.mouseDown();
        eventSender.mouseMoveTo(span.offsetLeft + span.offsetWidth / 2, span.offsetTop + span.offsetHeight / 2);
        eventSender.mouseUp();
    }
    shouldBe('document.getSelection().rangeCount', '0');
</script>
