third_party/blink/web_tests/media/video-controls-start-selection-expected.txt

PASS document.getSelection().rangeCount is 0
PASS successfullyParsed is true

TEST COMPLETE


Start selection on video with controls

 Some text here
